LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/navigating-across-documents/navigation-unload-form-submit-expected.txt

CONSOLE MESSAGE: TypeError: parent.finishedLoading is not a function. (In 'parent.finishedLoading()', 'parent.finishedLoading' is undefined)


PASS Tests that navigation during an unload caused by a form submit does nothing
